+++ quam_builder/architecture/superconducting/custom_gates/__init__.py
@@ -1,6 +1,7 @@
-from quam_builder.architecture.superconducting.custom_gates import cross_resonance, cz
+from quam_builder.architecture.superconducting.custom_gates.fixed_transmon_pair.two_qubit_gates import CRGate
+from quam_builder.architecture.superconducting.custom_gates.flux_tunable_transmon_pair.two_qubit_gates import CZGate
 
 __all__ = [
-    *cross_resonance.__all__,
-    *cz.__all__,
+    CRGate,
+    CZGate,
 ]
